+++ skel/src/benchmarksnoisy.c
@@ -2024,7 +2024,7 @@ void initbenchmarksnoisy()
     rot2 = malloc(sizeof(double*) * DIM);
     linearTF = malloc(sizeof(double*) * DIM);
     peaks21 = malloc(sizeof(double) * DIM * NHIGHPEAKS21);
-    rperm21 = malloc(sizeof(int) * fmax(DIM, NHIGHPEAKS21 - 1));
+    rperm21 = malloc(sizeof(int) * (int)fmax(DIM, NHIGHPEAKS21 - 1));
     arrScales21 = malloc(sizeof(double*) * NHIGHPEAKS21);
     Xlocal21 = malloc(sizeof(double*) * DIM);
